feat: 즐겨찾기 삭제 예외

+++ src/main/java/nextstep/favorite/applicataion/FavoriteService.java
@@ -54,6 +54,10 @@ public class FavoriteService {
     }
 
     public void removeFavorite(Long memberId, Long favoriteId) {
+        if (!favoriteRepository.existsByIdAndMemberId(favoriteId, memberId)) {
+            throw new IllegalArgumentException();
+        }
+
         favoriteRepository.deleteByIdAndMemberId(favoriteId, memberId);
     }
 }
